Minor cleanup and comments
