+++ src/ImageSharp.Drawing/Processing/TextDrawingRun.cs
@@ -1,47 +1,38 @@
 // Copyright (c) Six Labors.
 // Licensed under the Apache License, Version 2.0.
 
-using System;
-using System.Collections.Generic;
-using System.Text;
 using SixLabors.Fonts;
-using SixLabors.ImageSharp.Drawing.Processing;
 
-namespace SixLabors.Fonts
+namespace SixLabors.ImageSharp.Drawing.Processing
 {
-    public class TextDrawingOptions : TextOptions
-    {
-        public TextDrawingOptions(Font font)
-            : base(font)
-        {
-        }
-
-        public TextDrawingOptions(TextDrawingOptions options)
-            : base(options)
-        {
-        }
-
-        public new IReadOnlyList<TextDrawingRun> TextRuns
-        {
-            get => (IReadOnlyList<TextDrawingRun>)base.TextRuns;
-            set => base.TextRuns = value;
-        }
-    }
-
+    /// <summary>
+    /// Represents a run of drawable text spanning a series of graphemes within a string.
+    /// </summary>
     public class TextDrawingRun : TextRun
     {
+        /// <summary>
+        /// Gets or sets the brush used for filling this run.
+        /// </summary>
         public IBrush Brush { get; set; }
 
+        /// <summary>
+        /// Gets or sets the pen used for outlining this run.
+        /// </summary>
         public IPen Pen { get; set; }
 
+        /// <summary>
+        /// Gets or sets the pen used for drawing strikeout features for this run.
+        /// </summary>
         public IPen StrikeoutPen { get; set; }
 
+        /// <summary>
+        /// Gets or sets the pen used for drawing underline features for this run.
+        /// </summary>
         public IPen UnderlinePen { get; set; }
 
+        /// <summary>
+        /// Gets or sets the pen used for drawing overline features for this run.
+        /// </summary>
         public IPen OverlinePen { get; set; }
-
-        public TextDrawingRun()
-        {
-        }
     }
 }
