+++ tests/ImageSharp.Drawing.Benchmarks/Drawing/DrawTextOutline.cs
@@ -5,7 +5,6 @@ using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Linq;
 using BenchmarkDotNet.Attributes;
-using SixLabors.Fonts;
 using SixLabors.ImageSharp.Drawing.Processing;
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing;
